requirements.txt
streamlit
langgraph
langchain
openai
faiss-cpu
pandas
numpy
python-dotenv
langchain-openai
langchain-chroma 
chromadb